Edit footer formatting and create left container
The div for the logo container needed to be created so the footer was
moved into a left container with the logo container. The footer was
made to stick at the bottom of the page once more and the logo container
was placed where desired on the left half of the page.